# Conflicts:
#	CHANGELOG.md
#	docs/readthedocs/models/input/thermal/thermalhouse.md
#	docs/uml/main/input/ThermalDatamodelConcept.puml
#	src/main/java/edu/ie3/datamodel/io/factory/input/ThermalHouseInputFactory.java
#	src/main/java/edu/ie3/datamodel/models/input/thermal/ThermalHouseInput.java
#	src/test/groovy/edu/ie3/datamodel/io/factory/input/ThermalHouseInputFactoryTest.groovy
#	src/test/groovy/edu/ie3/datamodel/models/input/thermal/ThermalHouseInputTest.groovy
#	src/test/groovy/edu/ie3/datamodel/utils/validation/ThermalValidationUtilsTest.groovy
#	src/test/resources/edu/ie3/datamodel/io/sink/_sql/input_entities.sql

+++ src/main/java/edu/ie3/datamodel/io/factory/input/ThermalHouseInputFactory.java
@@ -62,7 +62,7 @@ public class ThermalHouseInputFactory
     final ComparableQuantity<Temperature> lowerTemperatureLimit =
         data.getQuantity(LOWER_TEMPERATURE_LIMIT, StandardUnits.TEMPERATURE);
     final String housingType = data.getField(HOUSING_TYPE);
-    final Integer numberInhabitants = data.getInt(NUMBER_INHABITANTS);
+    final double numberInhabitants = data.getDouble(NUMBER_INHABITANTS);
     return new ThermalHouseInput(
         uuid,
         id,
